refactor: migrate E2E tests to @vscode/test-cli
Replace manual runTest.ts + createMochaRunner.ts + per-suite index.ts
boilerplate with declarative .vscode-test.mjs configuration.
