--- packages/extension/tests/suite-multi/index.ts
@@ -1,3 +1,0 @@
-import { createMochaRunner } from '../createMochaRunner';
-
-export const run = createMochaRunner(__dirname);
