--- packages/extension/tests/suite/index.ts
@@ -1,3 +1,0 @@
-import { createMochaRunner } from '../createMochaRunner';
-
-export const run = createMochaRunner(__dirname);
